:wrench: CHORE: 테이블 업데이트 (#40)
