+++ src/main/java/shop/catchmind/picture/domain/Picture.java
@@ -1,10 +1,7 @@
 package shop.catchmind.picture.domain;
 
 import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import shop.catchmind.common.AuditingField;
 
 @Entity
@@ -17,35 +14,25 @@ public class Picture extends AuditingField {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "member_id")
-    private Long memberId;
-
-    @Column(name = "title", length = 15)
-    private String title;
-
     @Column(name = "image_url")
     private String imageUrl;
 
-    @Column(name = "result", length = 3000)
-    private String result;
+    @Column(name = "content", length = 3000)
+    private String content;
 
     @Column(name = "picture_type", nullable = false)
     @Enumerated(EnumType.STRING)
     private PictureType pictureType;
 
-    public void updateTitle(final String title) {
-        this.title = title;
-    }
-
-    public void updateResult(final String result) {
-        this.result = result;
-    }
+    @Setter
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "result_id")
+    private Result result;
 
     @Builder
-    public Picture(final Long memberId, final String title, final String imageUrl, final String result) {
-        this.memberId = memberId;
-        this.title = title;
+    private Picture(final String imageUrl, final String content, final PictureType pictureType) {
         this.imageUrl = imageUrl;
-        this.result = result;
+        this.content = content;
+        this.pictureType = pictureType;
     }
 }
