+++ src/main/resources/application.yml
@@ -109,7 +109,7 @@ spring:
       on-profile: local
   jpa:
     hibernate:
-      ddl-auto: validate
+      ddl-auto: create
 
 ---
 ### Develop Config
